:truck: Restructure templates into per-app layout
Refs #218
