+++ lib/templates/src/apps/debug/components/ErrorTitle.tsx
