+++ lib/templates/src/apps/debug/components/Section.tsx
