+++ lib/templates/src/apps/debug/components/Table/ArrayValue.tsx
