+++ lib/templates/src/apps/debug/components/Table/MapValue.tsx
