+++ lib/templates/src/apps/debug/components/Table/Row.tsx
