+++ lib/templates/src/apps/debug/components/Table/Table.tsx
