+++ lib/templates/src/apps/debug/components/Table/Value.tsx
@@ -1,4 +1,4 @@
-import { html } from '@/ui/lib/codecs'
+import { html } from '../../lib/codecs'
 
 export interface ValueProps {
   value: string | number | boolean
