+++ lib/templates/src/apps/debug/components/Table/index.ts
