+++ lib/templates/src/apps/debug/components/index.ts
@@ -1,0 +1,6 @@
+export { default as EnvironmentTable } from './EnvironmentTable'
+export { default as ErrorTitle } from './ErrorTitle'
+export { default as ErrorTraceback } from './ErrorTraceback'
+export { default as RequestTable } from './RequestTable'
+export { default as Section } from './Section'
+export { default as URLTree } from './URLTree'
