+++ lib/templates/src/apps/debug/data/Environment.ts
