+++ lib/templates/src/apps/debug/data/Error.ts
