+++ lib/templates/src/apps/debug/data/URLs.ts
