+++ lib/templates/src/apps/debug/data/index.ts
