+++ lib/templates/src/apps/schemas/components/Docs.tsx
@@ -1,9 +1,6 @@
 import { ApiReferenceReact } from '@scalar/api-reference-react'
 
-import { Docs as DocsData } from '@/data/schemas'
-
-import '@/ui/styles/docs.css'
-import '@scalar/api-reference-react/style.css'
+import { Docs as DocsData } from '../data'
 
 export default function Docs() {
   const docs = new DocsData()
