+++ lib/templates/src/apps/schemas/components/index.ts
