+++ lib/templates/src/apps/schemas/data/Docs.tsx
