+++ lib/templates/src/apps/schemas/data/index.ts
@@ -1,0 +1,1 @@
+export { default as Docs } from './Docs'
