+++ lib/templates/src/apps/schemas/styles/docs.css
