+++ lib/templates/src/ui/components/index.ts
@@ -1,9 +1,2 @@
-export { default as Docs } from './Docs'
-export { default as EnvironmentTable } from './EnvironmentTable'
-export { default as ErrorTitle } from './ErrorTitle'
-export { default as ErrorTraceback } from './ErrorTraceback'
 export { default as Footer } from './Footer'
 export { default as Menu } from './Menu'
-export { default as RequestTable } from './RequestTable'
-export { default as Section } from './Section'
-export { default as URLTree } from './URLTree'
